[#37] Refactoring Types and Separate in Modules
