+++ src/main/java/br/com/ppm/commons/type/Wrapper.java
@@ -1,0 +1,7 @@
+/*
+ * Copyright (C) 2020.
+ */
+package br.com.ppm.commons.type;
+
+public class Wrapper {
+}
