+++ src/main/java/br/com/ppm/commons/type/Wrappers.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright (C) 2020.
+ */
+
+package br.com.ppm.commons.type;
+
+import java.util.Map;
+
+public class Wrappers implements Map<String, Class<?>> {
+}
